+++ runtimes/server-interface/index.ts
@@ -8,4 +8,4 @@ export { Chat } from './chat'
 export * from './runtime'
 export * from './identity-management'
 export * from './notification'
-export * from './sdk-decorator'
+export * from './sdk-initializator'
